+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Added
 
-- Added Detekt task `detekt[Variant]All` with type resolution
+- Added Detekt tasks `detekt[Variant]All` with type resolution
 
 ## [0.11] (2021-07-06)